Begin making http api for socket events
